refactor(components): 拆分 BookCard 和 TTS 组件样式与逻辑
