+++ packages/app-expo/src/components/library/BookCard.tsx
@@ -1,5 +1,5 @@
-import { ClockIcon, DatabaseIcon, HashIcon, Loader2Icon, Trash2Icon } from "@/components/ui/Icon";
-import { type ThemeColors, fontSize, fontWeight, radius, useColors } from "@/styles/theme";
+import { ClockIcon, DatabaseIcon, Loader2Icon } from "@/components/ui/Icon";
+import { useColors } from "@/styles/theme";
 import { getPlatformService } from "@readany/core/services";
 /**
  * BookCard — Touch-optimized book card matching Tauri mobile MobileBookCard exactly.
@@ -10,16 +10,14 @@ import { memo, useEffect, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
 import {
   Animated,
-  Dimensions,
   Easing,
   Image,
-  Modal,
-  Pressable,
-  StyleSheet,
   Text,
   TouchableOpacity,
   View,
 } from "react-native";
+import { BookCardActionSheet } from "./BookCardActionSheet";
+import { makeStyles } from "./book-card-styles";
 
 const AnimatedLoader = () => {
   const spinValue = useRef(new Animated.Value(0)).current;
@@ -46,13 +44,6 @@ const AnimatedLoader = () => {
     </Animated.View>
   );
 };
-
-const SCREEN_PADDING = 16;
-const NUM_COLUMNS = 3;
-const GRID_GAP = 12;
-const screenWidth = Dimensions.get("window").width;
-const coverWidth = (screenWidth - SCREEN_PADDING * 2 - GRID_GAP * (NUM_COLUMNS - 1)) / NUM_COLUMNS;
-const coverHeight = coverWidth * (41 / 28);
 
 interface BookCardProps {
   book: Book;
@@ -279,429 +270,15 @@ export const BookCard = memo(function BookCard({
         </View>
       </TouchableOpacity>
 
-      {/* Action Sheet (long-press menu) — matches Tauri exactly */}
-      <Modal
+      <BookCardActionSheet
         visible={showActions}
-        transparent
-        animationType="slide"
-        onRequestClose={() => setShowActions(false)}
-      >
-        <Pressable style={s.overlay} onPress={() => setShowActions(false)} />
-        <View style={s.actionSheet}>
-          {/* Handle bar */}
-          <View style={s.actionHandle} />
-
-          {/* Book info header */}
-          <View style={s.actionHeader}>
-            <Text style={s.actionTitle} numberOfLines={1}>
-              {book.meta.title}
-            </Text>
-            {book.meta.author ? <Text style={s.actionAuthor}>{book.meta.author}</Text> : null}
-          </View>
-          <View style={s.actionDivider} />
-
-          {/* Manage tags */}
-          {onManageTags && (
-            <TouchableOpacity
-              style={s.actionItem}
-              onPress={() => {
-                setShowActions(false);
-                onManageTags(book);
-              }}
-            >
-              <HashIcon size={20} color={colors.mutedForeground} />
-              <Text style={s.actionLabel}>{t("home.manageTags", "管理标签")}</Text>
-            </TouchableOpacity>
-          )}
-
-          {/* Vectorize */}
-          {onVectorize && (
-            <TouchableOpacity
-              style={s.actionItem}
-              onPress={() => {
-                setShowActions(false);
-                onVectorize(book);
-              }}
-            >
-              <DatabaseIcon size={20} color={colors.mutedForeground} />
-              <Text style={s.actionLabel}>
-                {book.isVectorized
-                  ? t("home.vec_reindex", "重新索引")
-                  : t("home.vec_vectorize", "向量化")}
-              </Text>
-            </TouchableOpacity>
-          )}
-
-          {/* Delete */}
-          <TouchableOpacity
-            style={s.actionItemDestructive}
-            onPress={() => {
-              setShowActions(false);
-              onDelete(book.id);
-            }}
-          >
-            <Trash2Icon size={20} color={colors.destructive} />
-            <Text style={s.actionLabelDestructive}>{t("common.remove", "删除")}</Text>
-          </TouchableOpacity>
-
-          <View style={s.actionDivider} />
-          {/* Cancel */}
-          <TouchableOpacity style={s.actionCancel} onPress={() => setShowActions(false)}>
-            <Text style={s.actionCancelText}>{t("common.cancel", "取消")}</Text>
-          </TouchableOpacity>
-        </View>
-      </Modal>
+        book={book}
+        onClose={() => setShowActions(false)}
+        onManageTags={onManageTags}
+        onVectorize={onVectorize}
+        onDelete={onDelete}
+      />
     </>
   );
 });
 
-const makeStyles = (colors: ThemeColors) =>
-  StyleSheet.create({
-    container: { width: coverWidth },
-    coverWrap: {
-      width: coverWidth,
-      height: coverHeight,
-      borderRadius: radius.sm,
-      overflow: "hidden",
-      position: "relative",
-      // Book cover shadow
-      shadowColor: "#000",
-      shadowOffset: { width: 0, height: 2 },
-      shadowOpacity: 0.15,
-      shadowRadius: 4,
-      elevation: 3,
-    },
-    coverImage: {
-      width: "100%",
-      height: "100%",
-      borderRadius: radius.sm,
-    },
-    // Book spine crease effect — simulates desktop .book-spine linear-gradient overlay
-    spineOverlay: {
-      position: "absolute",
-      top: 0,
-      left: 0,
-      bottom: 0,
-      width: "8%",
-      flexDirection: "row",
-      zIndex: 2,
-    },
-    spineStrip1: {
-      width: "6%",
-      height: "100%",
-      backgroundColor: "rgba(0,0,0,0.10)",
-    },
-    spineStrip2: {
-      width: "8%",
-      height: "100%",
-      backgroundColor: "rgba(20,20,20,0.20)",
-    },
-    spineStrip3: {
-      width: "5%",
-      height: "100%",
-      backgroundColor: "rgba(240,240,240,0.40)",
-    },
-    spineStrip4: {
-      width: "18%",
-      height: "100%",
-      backgroundColor: "rgba(215,215,215,0.35)",
-    },
-    spineStrip5: {
-      width: "12%",
-      height: "100%",
-      backgroundColor: "rgba(150,150,150,0.25)",
-    },
-    spineStrip6: {
-      width: "20%",
-      height: "100%",
-      backgroundColor: "rgba(100,100,100,0.18)",
-    },
-    spineStrip7: {
-      width: "31%",
-      height: "100%",
-      backgroundColor: "rgba(175,175,175,0.12)",
-    },
-    spineEdgeRight: {
-      position: "absolute",
-      top: 0,
-      right: -coverWidth * 0.92,
-      bottom: 0,
-      width: coverWidth * 0.02,
-      backgroundColor: "rgba(30,30,30,0.12)",
-    },
-    spineTopHighlight: {
-      position: "absolute",
-      top: 0,
-      left: 0,
-      right: 0,
-      height: "3%",
-      backgroundColor: "rgba(240,240,240,0.15)",
-      zIndex: 3,
-    },
-    spineBottomShadow: {
-      position: "absolute",
-      bottom: 0,
-      left: 0,
-      right: 0,
-      height: "8%",
-      backgroundColor: "rgba(15,15,15,0.15)",
-      zIndex: 3,
-    },
-    fallbackCover: {
-      flex: 1,
-      borderRadius: radius.sm,
-      overflow: "hidden",
-      position: "relative",
-    },
-    fallbackGradientTop: {
-      position: "absolute",
-      top: 0,
-      left: 0,
-      right: 0,
-      height: "50%",
-      backgroundColor: colors.stone100,
-    },
-    fallbackGradientBottom: {
-      position: "absolute",
-      bottom: 0,
-      left: 0,
-      right: 0,
-      height: "50%",
-      backgroundColor: colors.stone200,
-    },
-    fallbackContentOverlay: {
-      flex: 1,
-      padding: 10,
-      alignItems: "center",
-      justifyContent: "center",
-      zIndex: 1,
-    },
-    fallbackTitleWrap: {
-      flex: 1,
-      alignItems: "center",
-      justifyContent: "center",
-    },
-    fallbackTitle: {
-      textAlign: "center",
-      fontSize: fontSize.sm,
-      fontWeight: fontWeight.medium,
-      fontFamily: "serif",
-      color: colors.stone500,
-      lineHeight: 18,
-    },
-    fallbackDivider: {
-      width: 32,
-      height: 1,
-      backgroundColor: `${colors.stone300}99`,
-      marginVertical: 6,
-    },
-    fallbackAuthorWrap: {
-      height: "25%",
-      alignItems: "center",
-      justifyContent: "center",
-    },
-    fallbackAuthor: {
-      textAlign: "center",
-      fontSize: 12,
-      fontFamily: "serif",
-      color: colors.stone400,
-    },
-    progressBarBg: {
-      position: "absolute",
-      bottom: 0,
-      left: 0,
-      right: 0,
-      height: 2,
-      backgroundColor: "rgba(0,0,0,0.1)",
-    },
-    progressBarFill: {
-      height: 2,
-      backgroundColor: colors.primary,
-      opacity: 0.8,
-    },
-    // Vectorization overlay
-    vecOverlay: {
-      position: "absolute",
-      top: 0,
-      left: 0,
-      right: 0,
-      bottom: 0,
-      backgroundColor: "rgba(0,0,0,0.5)",
-      borderRadius: radius.sm,
-      alignItems: "center",
-      justifyContent: "center",
-    },
-    vecOverlayText: {
-      marginTop: 6,
-      fontSize: 14,
-      fontWeight: fontWeight.medium,
-      color: "#fff",
-    },
-    queuedOverlay: {
-      position: "absolute",
-      top: 0,
-      left: 0,
-      right: 0,
-      bottom: 0,
-      backgroundColor: "rgba(0,0,0,0.35)",
-      borderRadius: radius.sm,
-      alignItems: "center",
-      justifyContent: "center",
-    },
-    queuedOverlayText: {
-      marginTop: 4,
-      fontSize: 12,
-      fontWeight: fontWeight.medium,
-      color: "#fff",
-    },
-    // Remote status overlay (on-demand download)
-    remoteOverlay: {
-      position: "absolute",
-      top: 0,
-      left: 0,
-      right: 0,
-      bottom: 0,
-      backgroundColor: "rgba(59, 130, 246, 0.6)", // Blue tint
-      borderRadius: radius.sm,
-      alignItems: "center",
-      justifyContent: "center",
-    },
-    remoteOverlayText: {
-      fontSize: 12,
-      fontWeight: fontWeight.medium,
-      color: "#fff",
-      backgroundColor: "rgba(0,0,0,0.4)",
-      paddingHorizontal: 8,
-      paddingVertical: 4,
-      borderRadius: radius.sm,
-    },
-    // Downloading overlay
-    downloadingOverlay: {
-      position: "absolute",
-      top: 0,
-      left: 0,
-      right: 0,
-      bottom: 0,
-      backgroundColor: "rgba(0,0,0,0.5)",
-      borderRadius: radius.sm,
-      alignItems: "center",
-      justifyContent: "center",
-    },
-    downloadingOverlayText: {
-      marginTop: 6,
-      fontSize: 14,
-      fontWeight: fontWeight.medium,
-      color: "#fff",
-    },
-    vecBadge: {
-      position: "absolute",
-      top: 2,
-      left: 2,
-      flexDirection: "row",
-      alignItems: "center",
-      gap: 2,
-      backgroundColor: "rgba(22,163,74,0.8)",
-      borderRadius: radius.sm,
-      paddingHorizontal: 4,
-      paddingVertical: 2,
-    },
-    vecBadgeText: { fontSize: 7, fontWeight: fontWeight.medium, color: "#fff" },
-    infoWrap: { paddingTop: 6, paddingHorizontal: 1 },
-    bookTitle: {
-      fontSize: 13,
-      fontWeight: fontWeight.semibold,
-      color: colors.foreground,
-      lineHeight: 14,
-    },
-    tagsRow: { flexDirection: "row", flexWrap: "wrap", gap: 3, marginTop: 3 },
-    tagBadge: {
-      backgroundColor: `${colors.muted}`,
-      borderRadius: radius.full,
-      paddingHorizontal: 6,
-      paddingVertical: 1,
-    },
-    tagText: { fontSize: 8, color: colors.mutedForeground },
-    tagBadgeUncategorized: {
-      backgroundColor: `${colors.muted}80`,
-      borderRadius: radius.full,
-      paddingHorizontal: 6,
-      paddingVertical: 1,
-    },
-    tagTextUncategorized: { fontSize: 8, color: `${colors.mutedForeground}99` },
-    tagOverflow: { fontSize: 8, color: `${colors.mutedForeground}99`, alignSelf: "center" },
-    statusRow: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-      marginTop: 3,
-      minHeight: 14,
-    },
-    progressText: { fontSize: 9, color: colors.mutedForeground, fontVariant: ["tabular-nums"] },
-    completeText: { fontSize: 9, fontWeight: fontWeight.medium, color: "#16a34a" },
-    newBadge: {
-      backgroundColor: `${colors.primary}14`,
-      borderRadius: radius.full,
-      paddingHorizontal: 5,
-      paddingVertical: 1,
-    },
-    newText: { fontSize: 8, fontWeight: fontWeight.medium, color: colors.primary },
-    formatText: {
-      fontSize: 8,
-      color: `${colors.mutedForeground}99`,
-      textTransform: "uppercase",
-      letterSpacing: 0.5,
-    },
-    // Action Sheet
-    overlay: {
-      flex: 1,
-      backgroundColor: "rgba(0,0,0,0.4)",
-    },
-    actionSheet: {
-      backgroundColor: colors.background,
-      borderTopLeftRadius: 16,
-      borderTopRightRadius: 16,
-      paddingBottom: 34,
-    },
-    actionHandle: {
-      width: 40,
-      height: 4,
-      borderRadius: 2,
-      backgroundColor: colors.muted,
-      alignSelf: "center",
-      marginTop: 12,
-      marginBottom: 8,
-    },
-    actionHeader: { paddingHorizontal: 20, paddingBottom: 12 },
-    actionTitle: {
-      fontSize: fontSize.base,
-      fontWeight: fontWeight.semibold,
-      color: colors.foreground,
-    },
-    actionAuthor: { fontSize: fontSize.xs, color: colors.mutedForeground, marginTop: 2 },
-    actionDivider: { height: 0.5, backgroundColor: colors.border },
-    actionItem: {
-      flexDirection: "row",
-      alignItems: "center",
-      gap: 12,
-      paddingHorizontal: 20,
-      paddingVertical: 14,
-    },
-    actionLabel: { fontSize: fontSize.base, color: colors.foreground },
-    actionItemDestructive: {
-      flexDirection: "row",
-      alignItems: "center",
-      gap: 12,
-      paddingHorizontal: 20,
-      paddingVertical: 14,
-    },
-    actionLabelDestructive: { fontSize: fontSize.base, color: colors.destructive },
-    actionCancel: {
-      alignItems: "center",
-      paddingVertical: 14,
-    },
-    actionCancelText: {
-      fontSize: fontSize.base,
-      fontWeight: fontWeight.medium,
-      color: colors.foreground,
-    },
-  });
