+++ packages/app-expo/src/components/tts/FloatingTTSBubble.tsx
@@ -3,334 +3,23 @@
  *
  * Rendered as a sibling to NavigationContainer in AppInner so it floats
  * above every screen. Tapping it expands a compact player modal.
- *
- * Design:
- * - Bubble always shows headphones icon (no play/pause switching)
- * - Pulsing ring animation when playing; no animation when paused
- * - No X close badge — stop is inside the expanded mini player
- * - Tap bubble → toggle mini player
- * - Long press bubble → stop TTS
  */
 import { useTTSStore } from "@/stores";
-import {
-  BookOpenIcon,
-  HeadphonesIcon,
-  PauseIcon,
-  PlayIcon,
-  ScrollTextIcon,
-  SquareIcon,
-} from "@/components/ui/Icon";
-import { useReaderStore } from "@/stores/reader-store";
-import { fontSize, radius, useColors } from "@/styles/theme";
-import { pushRoute } from "@/lib/navigationRef";
-import { eventBus } from "@readany/core/utils/event-bus";
+import { HeadphonesIcon } from "@/components/ui/Icon";
+import { useColors } from "@/styles/theme";
+import { TTSMiniPlayer } from "./TTSMiniPlayer";
 import { useCallback, useEffect, useRef, useState } from "react";
-import { useTranslation } from "react-i18next";
 import {
   ActivityIndicator,
   Animated,
   Dimensions,
   Easing,
-  LayoutChangeEvent,
-  Modal,
-  PanResponder,
-  Pressable,
   StyleSheet,
-  Text,
   TouchableOpacity,
-  View,
 } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 
-// ─── Compact expanded player modal ───────────────────────────────────────────
-
-function TTSMiniPlayer({
-  visible,
-  onClose,
-  anchorLayout,
-}: {
-  visible: boolean;
-  onClose: () => void;
-  anchorLayout: { left: number; top: number; size: number; screenWidth: number; screenHeight: number } | null;
-}) {
-  const { t } = useTranslation();
-  const colors = useColors();
-  const insets = useSafeAreaInsets();
-
-  const playState = useTTSStore((s) => s.playState);
-  const currentBookTitle = useTTSStore((s) => s.currentBookTitle);
-  const currentChapterTitle = useTTSStore((s) => s.currentChapterTitle);
-  const currentBookId = useTTSStore((s) => s.currentBookId);
-  const currentLocationCfi = useTTSStore((s) => s.currentLocationCfi);
-  const goToCfiFn = useReaderStore((s) => s.goToCfiFn);
-  const config = useTTSStore((s) => s.config);
-  const pause = useTTSStore((s) => s.pause);
-  const resume = useTTSStore((s) => s.resume);
-  const stop = useTTSStore((s) => s.stop);
-  const updateConfig = useTTSStore((s) => s.updateConfig);
-
-  const handleStop = useCallback(() => {
-    stop();
-    onClose();
-  }, [stop, onClose]);
-
-  const handleJumpToCurrentLocation = useCallback(() => {
-    let handled = false;
-    if (currentBookId && currentLocationCfi) {
-      eventBus.emit("tts:jump-to-current", {
-        bookId: currentBookId,
-        cfi: currentLocationCfi,
-        respond: () => {
-          handled = true;
-        },
-      });
-    }
-    if (!handled && currentLocationCfi && goToCfiFn) {
-      goToCfiFn(currentLocationCfi);
-      onClose();
-      return;
-    }
-    if (!handled && currentBookId) {
-      pushRoute("Reader", { bookId: currentBookId, cfi: currentLocationCfi || undefined });
-    }
-    onClose();
-  }, [currentBookId, currentLocationCfi, goToCfiFn, onClose]);
-
-  const handleOpenLyricsPage = useCallback(() => {
-    if (!currentBookId) return;
-    let handled = false;
-    eventBus.emit("tts:open-lyrics-page", {
-      bookId: currentBookId,
-      respond: () => {
-        handled = true;
-      },
-    });
-    if (!handled) {
-      pushRoute("Reader", { bookId: currentBookId, openTTS: true });
-    }
-    onClose();
-  }, [currentBookId, onClose]);
-
-  const handlePlayPause = useCallback(() => {
-    if (playState === "playing") {
-      pause();
-    } else if (playState === "paused") {
-      resume();
-    }
-  }, [playState, pause, resume]);
-
-  const adjustRate = useCallback(
-    (delta: number) => {
-      const newRate = Math.round(Math.max(0.5, Math.min(2.0, config.rate + delta)) * 10) / 10;
-      updateConfig({ rate: newRate });
-    },
-    [config.rate, updateConfig],
-  );
-
-  // Pulse animation for the headphones icon when playing
-  const pulseAnim = useRef(new Animated.Value(1)).current;
-
-  useEffect(() => {
-    if (playState === "playing") {
-      const anim = Animated.loop(
-        Animated.sequence([
-          Animated.timing(pulseAnim, {
-            toValue: 1.15,
-            duration: 600,
-            easing: Easing.inOut(Easing.ease),
-            useNativeDriver: true,
-          }),
-          Animated.timing(pulseAnim, {
-            toValue: 1,
-            duration: 600,
-            easing: Easing.inOut(Easing.ease),
-            useNativeDriver: true,
-          }),
-        ]),
-      );
-      anim.start();
-      return () => anim.stop();
-    }
-    pulseAnim.setValue(1);
-  }, [playState, pulseAnim]);
-
-  const statusText =
-    playState === "loading"
-      ? "加载中…"
-      : playState === "playing"
-        ? "播放中"
-        : playState === "paused"
-          ? "已暂停"
-          : "已停止";
-
-  const panelWidth = Math.min(388, Math.max(320, (anchorLayout?.screenWidth || 360) - 16));
-  const [panelHeight, setPanelHeight] = useState(152);
-  const [panelMeasured, setPanelMeasured] = useState(false);
-  const clamp = (value: number, min: number, max: number) => Math.min(Math.max(value, min), max);
-  const anchor = anchorLayout ?? {
-    left: 16,
-    top: 120,
-    size: BUBBLE_SIZE,
-    screenWidth: Dimensions.get("window").width,
-    screenHeight: Dimensions.get("window").height,
-  };
-  const left = clamp(
-    anchor.left + anchor.size / 2 - panelWidth / 2,
-    10,
-    anchor.screenWidth - panelWidth - 10,
-  );
-  const aboveGap = 10;
-  const belowGap = 10;
-  const safeTop = (insets.top || 12) + 8;
-  const safeBottom = anchor.screenHeight - panelHeight - Math.max(insets.bottom, 16) - 8;
-  const aboveTop = anchor.top - panelHeight - aboveGap;
-  const belowTop = anchor.top + anchor.size + belowGap;
-  const canPlaceAbove = aboveTop >= safeTop;
-  const canPlaceBelow = belowTop <= safeBottom;
-  const top = canPlaceAbove
-    ? aboveTop
-    : canPlaceBelow
-      ? belowTop
-      : clamp(belowTop, safeTop, safeBottom);
-
-  const handlePanelLayout = useCallback((event: LayoutChangeEvent) => {
-    const nextHeight = Math.ceil(event.nativeEvent.layout.height || 0);
-    if (nextHeight > 0) {
-      if (nextHeight !== panelHeight) {
-        setPanelHeight(nextHeight);
-      }
-      if (!panelMeasured) {
-        setPanelMeasured(true);
-      }
-    }
-  }, [panelHeight, panelMeasured]);
-
-  return (
-    <Modal visible={visible} transparent animationType="fade" onRequestClose={onClose}>
-      <Pressable
-        style={StyleSheet.absoluteFillObject}
-        onPress={onClose}
-        accessible={false}
-      />
-      <View
-        style={[
-          styles.miniPlayerContainer,
-          {
-            backgroundColor: colors.card,
-            borderColor: colors.border,
-            left,
-            top,
-            width: panelWidth,
-            opacity: panelMeasured || !visible ? 1 : 0,
-          },
-        ]}
-        pointerEvents="box-none"
-        onLayout={handlePanelLayout}
-      >
-        {/* Header row */}
-        <View style={styles.miniPlayerHeader}>
-          <Animated.View style={{ transform: [{ scale: pulseAnim }] }}>
-            <HeadphonesIcon size={18} color={colors.primary} />
-          </Animated.View>
-          <View style={styles.miniPlayerTitleGroup}>
-            <Text style={[styles.miniPlayerBook, { color: colors.foreground }]} numberOfLines={1}>
-              {currentBookTitle || "正在听书"}
-            </Text>
-            {!!currentChapterTitle && (
-              <Text
-                style={[styles.miniPlayerChapter, { color: colors.mutedForeground }]}
-                numberOfLines={1}
-              >
-                {currentChapterTitle}
-              </Text>
-            )}
-          </View>
-          <Text style={[styles.miniPlayerStatus, { color: colors.mutedForeground }]}>
-            {statusText}
-          </Text>
-        </View>
-
-        {/* Divider */}
-        <View style={[styles.miniPlayerDivider, { backgroundColor: colors.border }]} />
-
-        {/* Controls row */}
-        <View style={styles.miniPlayerControls}>
-          {/* Rate adjust */}
-          <TouchableOpacity
-            style={[styles.miniPlayerRateBtn, { backgroundColor: colors.muted }]}
-            onPress={() => adjustRate(-0.1)}
-          >
-            <Text style={[styles.miniPlayerRateText, { color: colors.foreground }]}>−</Text>
-          </TouchableOpacity>
-
-          <Text style={[styles.miniPlayerRateValue, { color: colors.mutedForeground }]}>
-            {config.rate.toFixed(1)}x
-          </Text>
-
-          <TouchableOpacity
-            style={[styles.miniPlayerRateBtn, { backgroundColor: colors.muted }]}
-            onPress={() => adjustRate(0.1)}
-          >
-            <Text style={[styles.miniPlayerRateText, { color: colors.foreground }]}>+</Text>
-          </TouchableOpacity>
-
-          <View style={[styles.miniPlayerDividerV, { backgroundColor: colors.border }]} />
-
-          {/* Play/Pause */}
-          <TouchableOpacity
-            style={[styles.miniPlayerPlayBtn, { backgroundColor: colors.primary }]}
-            onPress={handlePlayPause}
-            disabled={playState === "loading" || playState === "stopped"}
-          >
-            {playState === "loading" ? (
-              <ActivityIndicator size="small" color={colors.primaryForeground} />
-            ) : playState === "playing" ? (
-              <PauseIcon size={20} color={colors.primaryForeground} />
-            ) : (
-              <PlayIcon size={20} color={colors.primaryForeground} />
-            )}
-          </TouchableOpacity>
-
-          {/* Stop */}
-          <TouchableOpacity
-            style={[styles.miniPlayerStopBtn, { backgroundColor: colors.muted }]}
-            onPress={handleStop}
-            accessibilityRole="button"
-            accessibilityLabel={t("tts.stop", "停止")}
-          >
-            <SquareIcon size={16} color={colors.foreground} />
-          </TouchableOpacity>
-
-          {!!currentBookId && <View style={[styles.miniPlayerDividerV, { backgroundColor: colors.border }]} />}
-
-          {!!currentBookId && (
-            <TouchableOpacity
-              style={[styles.miniPlayerGoBtn, { backgroundColor: colors.muted }]}
-              onPress={handleJumpToCurrentLocation}
-              accessibilityRole="button"
-              accessibilityLabel={t("tts.jumpToCurrentLocation")}
-            >
-              <BookOpenIcon size={16} color={colors.foreground} />
-            </TouchableOpacity>
-          )}
-
-          {!!currentBookId && (
-            <TouchableOpacity
-              style={[styles.miniPlayerGoBtn, { backgroundColor: colors.muted }]}
-              onPress={handleOpenLyricsPage}
-              accessibilityRole="button"
-              accessibilityLabel={t("tts.openLyricsPage", "跳到歌词页")}
-            >
-              <ScrollTextIcon size={16} color={colors.foreground} />
-            </TouchableOpacity>
-          )}
-        </View>
-      </View>
-    </Modal>
-  );
-}
-
-// ─── Main floating bubble ─────────────────────────────────────────────────────
+const BUBBLE_SIZE = 56;
 
 export function FloatingTTSBubble() {
   const colors = useColors();
@@ -349,47 +38,34 @@ export function FloatingTTSBubble() {
     screenHeight: number;
   } | null>(null);
 
-  // Only show when TTS is active (playing or paused)
   const isActive = playState === "playing" || playState === "paused" || playState === "loading";
 
-  // When TTS stops, close the mini player
   useEffect(() => {
-    if (!isActive) {
-      setShowPlayer(false);
-    }
+    if (!isActive) setShowPlayer(false);
   }, [isActive]);
 
-  // Draggable position — starts at bottom right
   const pan = useRef(new Animated.ValueXY({ x: 0, y: 0 })).current;
   const bubbleRight = useRef(20);
   const bubbleBottom = useRef(120);
-  const bubbleRef = useRef<View>(null);
+  const bubbleRef = useRef<Animated.LegacyRef<typeof Animated.View>>(null);
 
   const panResponder = useRef(
-    PanResponder.create({
+    require("react-native").PanResponder.create({
       onStartShouldSetPanResponder: () => true,
-      onMoveShouldSetPanResponder: (_, gs) =>
+      onMoveShouldSetPanResponder: (_: unknown, gs: { dx: number; dy: number }) =>
         Math.abs(gs.dx) > 5 || Math.abs(gs.dy) > 5,
       onPanResponderGrant: () => {
         pan.setOffset({ x: (pan.x as any)._value, y: (pan.y as any)._value });
         pan.setValue({ x: 0, y: 0 });
       },
-      onPanResponderMove: Animated.event([null, { dx: pan.x, dy: pan.y }], {
-        useNativeDriver: false,
-      }),
-      onPanResponderRelease: () => {
-        pan.flattenOffset();
-      },
+      onPanResponderMove: Animated.event([null, { dx: pan.x, dy: pan.y }], { useNativeDriver: false }),
+      onPanResponderRelease: () => { pan.flattenOffset(); },
     }),
   ).current;
 
   useEffect(() => {
-    const id = pan.addListener((value) => {
-      setBubbleOffset({ x: value.x, y: value.y });
-    });
-    return () => {
-      pan.removeListener(id);
-    };
+    const id = pan.addListener((value) => { setBubbleOffset({ x: value.x, y: value.y }); });
+    return () => { pan.removeListener(id); };
   }, [pan]);
 
   // Ripple pulse rings — only when playing
@@ -402,24 +78,15 @@ export function FloatingTTSBubble() {
         Animated.loop(
           Animated.sequence([
             Animated.delay(delay),
-            Animated.timing(anim, {
-              toValue: 1,
-              duration: 1600,
-              easing: Easing.out(Easing.ease),
-              useNativeDriver: true,
-            }),
+            Animated.timing(anim, { toValue: 1, duration: 1600, easing: Easing.out(Easing.ease), useNativeDriver: true }),
             Animated.timing(anim, { toValue: 0, duration: 0, useNativeDriver: true }),
           ]),
         );
-
       const a1 = makeRipple(ring1, 0);
       const a2 = makeRipple(ring2, 700);
       a1.start();
       a2.start();
-      return () => {
-        a1.stop();
-        a2.stop();
-      };
+      return () => { a1.stop(); a2.stop(); };
     }
     ring1.setValue(0);
     ring2.setValue(0);
@@ -430,40 +97,28 @@ export function FloatingTTSBubble() {
     opacity: anim.interpolate({ inputRange: [0, 0.3, 1], outputRange: [0, 0.35, 0] }),
   });
 
-  const handleBubbleTap = useCallback(() => {
-    setShowPlayer((v) => !v);
-  }, []);
+  const handleBubbleTap = useCallback(() => { setShowPlayer((v) => !v); }, []);
 
   const measureBubble = useCallback(() => {
     requestAnimationFrame(() => {
-      bubbleRef.current?.measureInWindow((left, top, width, height) => {
+      (bubbleRef.current as any)?.measureInWindow((left: number, top: number, width: number, height: number) => {
         if (width > 0 && height > 0) {
-          setAnchorLayout({
-            left,
-            top,
-            size: Math.max(width, height),
-            screenWidth,
-            screenHeight,
-          });
+          setAnchorLayout({ left, top, size: Math.max(width, height), screenWidth, screenHeight });
         }
       });
     });
   }, [screenHeight, screenWidth]);
 
   useEffect(() => {
-    if (!isActive) {
-      setAnchorLayout(null);
-      return;
-    }
+    if (!isActive) { setAnchorLayout(null); return; }
     measureBubble();
   }, [bubbleOffset.x, bubbleOffset.y, isActive, measureBubble, showPlayer]);
 
   return (
     <>
-      {/* Draggable bubble — only rendered when active */}
       {isActive && (
         <Animated.View
-          ref={bubbleRef}
+          ref={bubbleRef as any}
           collapsable={false}
           style={[
             styles.bubbleWrapper,
@@ -475,7 +130,6 @@ export function FloatingTTSBubble() {
           ]}
           {...panResponder.panHandlers}
         >
-          {/* Ripple rings — playing only */}
           <Animated.View
             style={[styles.bubbleRing, { backgroundColor: colors.primary }, makeRingStyle(ring1)]}
             pointerEvents="none"
@@ -484,8 +138,6 @@ export function FloatingTTSBubble() {
             style={[styles.bubbleRing, { backgroundColor: colors.primary }, makeRingStyle(ring2)]}
             pointerEvents="none"
           />
-
-          {/* Main bubble */}
           <TouchableOpacity
             style={[styles.bubble, { backgroundColor: colors.primary }]}
             onPress={handleBubbleTap}
@@ -500,8 +152,6 @@ export function FloatingTTSBubble() {
         </Animated.View>
       )}
 
-      {/* Mini player modal — visible prop controls it independently so it
-          can animate out even as isActive drops to false */}
       <TTSMiniPlayer
         visible={showPlayer && isActive}
         onClose={() => setShowPlayer(false)}
@@ -510,10 +160,6 @@ export function FloatingTTSBubble() {
     </>
   );
 }
-
-// ─── Styles ───────────────────────────────────────────────────────────────────
-
-const BUBBLE_SIZE = 56;
 
 const styles = StyleSheet.create({
   bubbleWrapper: {
@@ -541,92 +187,5 @@ const styles = StyleSheet.create({
     shadowOpacity: 0.25,
     shadowRadius: 8,
     elevation: 12,
-  },
-  // Mini player
-  miniPlayerContainer: {
-    position: "absolute",
-    borderRadius: radius.xl,
-    borderWidth: 1,
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 8 },
-    shadowOpacity: 0.18,
-    shadowRadius: 16,
-    elevation: 20,
-    overflow: "hidden",
-  },
-  miniPlayerHeader: {
-    flexDirection: "row",
-    alignItems: "center",
-    paddingHorizontal: 16,
-    paddingVertical: 14,
-    gap: 10,
-  },
-  miniPlayerTitleGroup: {
-    flex: 1,
-    gap: 2,
-  },
-  miniPlayerBook: {
-    fontSize: fontSize.sm,
-    fontWeight: "600",
-  },
-  miniPlayerChapter: {
-    fontSize: fontSize.xs,
-  },
-  miniPlayerStatus: {
-    fontSize: fontSize.xs,
-  },
-  miniPlayerDivider: {
-    height: StyleSheet.hairlineWidth,
-    marginHorizontal: 16,
-  },
-  miniPlayerControls: {
-    flexDirection: "row",
-    alignItems: "center",
-    paddingHorizontal: 14,
-    paddingVertical: 12,
-    gap: 6,
-  },
-  miniPlayerRateBtn: {
-    width: 32,
-    height: 32,
-    borderRadius: radius.md,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  miniPlayerRateText: {
-    fontSize: 18,
-    fontWeight: "500",
-    lineHeight: 20,
-  },
-  miniPlayerRateValue: {
-    fontSize: fontSize.xs,
-    width: 40,
-    textAlign: "center",
-  },
-  miniPlayerDividerV: {
-    width: StyleSheet.hairlineWidth,
-    height: 24,
-    marginHorizontal: 2,
-  },
-  miniPlayerPlayBtn: {
-    width: 44,
-    height: 44,
-    borderRadius: 22,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  miniPlayerStopBtn: {
-    width: 36,
-    height: 36,
-    borderRadius: radius.lg,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  miniPlayerGoBtn: {
-    width: 36,
-    height: 36,
-    borderRadius: radius.lg,
-    alignItems: "center",
-    justifyContent: "center",
   },
 });
