+++ packages/app-expo/src/screens/reader/reader-styles.ts
@@ -1,727 +1,16 @@
 /**
- * Styles for ReaderScreen.
+ * Styles for ReaderScreen — aggregated from sub-modules.
+ * Keep this file as a single import point for backward compatibility.
  */
-import { Dimensions, StyleSheet } from "react-native";
-import {
-  type ThemeColors,
-  fontSize,
-  fontWeight,
-  radius,
-  withOpacity,
-} from "@/styles/theme";
+import type { ThemeColors } from "@/styles/theme";
+import { makeToolbarStyles } from "./styles/reader-base-styles";
+import { makeSheetStyles } from "./styles/reader-sheet-styles";
+import { makeNoteStyles } from "./styles/reader-note-styles";
 
-const SCREEN_HEIGHT = Dimensions.get("window").height;
+export { TOOLTIP_FG, TOOLTIP_MUTED, noteTooltipMdStyles } from "./styles/reader-note-styles";
 
-export const TOOLTIP_FG = "#f1f5f9";
-export const TOOLTIP_MUTED = "rgba(148, 163, 184, 0.5)";
-
-export const noteTooltipMdStyles = {
-  body: { color: TOOLTIP_FG, fontSize: 13, lineHeight: 19 },
-  textgroup: { color: TOOLTIP_FG, fontSize: 13, lineHeight: 19 },
-  text: { color: TOOLTIP_FG, fontSize: 13, lineHeight: 19 },
-  paragraph: { color: TOOLTIP_FG, fontSize: 13, lineHeight: 19, marginBottom: 4, marginTop: 0 },
-  heading1: {
-    color: "#fff",
-    fontSize: 15,
-    fontWeight: "600" as const,
-    marginBottom: 4,
-    marginTop: 4,
-  },
-  heading2: {
-    color: "#fff",
-    fontSize: 14,
-    fontWeight: "600" as const,
-    marginBottom: 3,
-    marginTop: 3,
-  },
-  heading3: {
-    color: "#fff",
-    fontSize: 13,
-    fontWeight: "600" as const,
-    marginBottom: 2,
-    marginTop: 2,
-  },
-  strong: { fontWeight: "700" as const, color: "#fff" },
-  em: { fontStyle: "italic" as const, color: "#e2e8f0" },
-  link: { color: "#60a5fa" },
-  code_inline: {
-    backgroundColor: "rgba(255,255,255,0.1)",
-    color: TOOLTIP_FG,
-    fontSize: 14,
-    fontFamily: "Menlo",
-  },
-  code_block: {
-    backgroundColor: "rgba(0,0,0,0.3)",
-    color: TOOLTIP_FG,
-    fontSize: 14,
-    fontFamily: "Menlo",
-    padding: 8,
-  },
-  fence: {
-    backgroundColor: "rgba(0,0,0,0.3)",
-    color: TOOLTIP_FG,
-    fontSize: 14,
-    fontFamily: "Menlo",
-    padding: 8,
-  },
-  blockquote: {
-    borderLeftWidth: 2,
-    borderLeftColor: TOOLTIP_MUTED,
-    paddingLeft: 10,
-    backgroundColor: "transparent",
-    color: TOOLTIP_FG,
-  },
-  bullet_list: { marginVertical: 2 },
-  ordered_list: { marginVertical: 2 },
-  list_item: { marginBottom: 2, flexDirection: "row" as const },
-  bullet_list_icon: { color: TOOLTIP_FG, marginLeft: 0, marginRight: 8 },
-  bullet_list_content: { color: TOOLTIP_FG, flex: 1 },
-  ordered_list_icon: { color: TOOLTIP_FG, marginLeft: 0, marginRight: 8 },
-  ordered_list_content: { color: TOOLTIP_FG, flex: 1 },
-  hardbreak: { color: TOOLTIP_FG },
-  softbreak: { color: TOOLTIP_FG },
-};
-
-export const makeStyles = (colors: ThemeColors) =>
-  StyleSheet.create({
-    container: { flex: 1, backgroundColor: colors.background },
-    readerStage: { flex: 1 },
-    webview: { flex: 1 },
-    loadingWrap: { flex: 1, alignItems: "center", justifyContent: "center", gap: 12 },
-    loadingText: { fontSize: fontSize.sm, color: colors.mutedForeground },
-    errorText: {
-      fontSize: fontSize.base,
-      color: colors.destructive,
-      textAlign: "center",
-      paddingHorizontal: 24,
-    },
-    backButton: {
-      marginTop: 16,
-      paddingHorizontal: 24,
-      paddingVertical: 10,
-      borderRadius: radius.lg,
-      backgroundColor: colors.primary,
-    },
-    backButtonText: {
-      fontSize: fontSize.sm,
-      fontWeight: fontWeight.medium,
-      color: colors.primaryForeground,
-    },
-    loadingOverlay: {
-      position: "absolute",
-      top: 0,
-      left: 0,
-      right: 0,
-      bottom: 0,
-      alignItems: "center",
-      justifyContent: "center",
-      backgroundColor: colors.background,
-      zIndex: 20,
-    },
-
-    topToolbar: {
-      position: "absolute",
-      zIndex: 34,
-      left: 0,
-      right: 0,
-      top: 0,
-    },
-    topToolbarBar: {
-      backgroundColor: withOpacity(colors.background, 0.94),
-      borderBottomWidth: StyleSheet.hairlineWidth,
-      borderBottomColor: withOpacity(colors.foreground, 0.1),
-    },
-    topToolbarRow: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-    },
-    topToolbarBackBtn: {
-      width: 48,
-      height: 48,
-      borderRadius: 24,
-      alignItems: "center",
-      justifyContent: "center",
-    },
-    topToolbarSideSlot: {
-      width: 68,
-      justifyContent: "center",
-      alignItems: "flex-start",
-    },
-    topToolbarSpacer: {
-      flex: 1,
-      minHeight: 44,
-    },
-    topToolbarTitleWrap: {
-      flex: 1,
-      minWidth: 0,
-      alignItems: "center",
-      justifyContent: "center",
-      paddingHorizontal: 12,
-    },
-    topToolbarTitleText: {
-      fontSize: fontSize.sm,
-      color: colors.foreground,
-      fontWeight: fontWeight.medium,
-    },
-    topToolbarMetaWrap: {
-      width: 60,
-      alignItems: "flex-end",
-      justifyContent: "center",
-    },
-    topToolbarMetaText: {
-      fontSize: fontSize.sm,
-      color: colors.mutedForeground,
-      fontWeight: fontWeight.medium,
-    },
-    topToolbarRight: {
-      flexDirection: "row",
-      alignItems: "center",
-      gap: 2,
-    },
-    topToolbarBtn: {
-      width: 40,
-      height: 40,
-      alignItems: "center",
-      justifyContent: "center",
-    },
-    topToolbarBtnActive: {
-      backgroundColor: withOpacity(colors.foreground, 0.06),
-    },
-    topToolbarProgressTrack: {
-      height: 2,
-      backgroundColor: withOpacity(colors.foreground, 0.08),
-      overflow: "hidden",
-    },
-    topToolbarProgressFill: {
-      height: "100%",
-      backgroundColor: withOpacity(colors.foreground, 0.48),
-    },
-
-    floatingTools: {
-      position: "absolute",
-      zIndex: 34,
-      gap: 10,
-      alignItems: "center",
-    },
-    floatingToolBtn: {
-      width: 52,
-      height: 52,
-      borderRadius: 26,
-      alignItems: "center",
-      justifyContent: "center",
-      backgroundColor: "rgba(76, 82, 94, 0.88)",
-    },
-    floatingToolBtnActive: {
-      backgroundColor: "rgba(59, 130, 246, 0.9)",
-    },
-
-    bottomToolbar: {
-      position: "absolute",
-      bottom: 0,
-      zIndex: 30,
-      left: 0,
-      right: 0,
-    },
-    bottomToolbarGlass: {
-      backgroundColor: withOpacity(colors.background, 0.98),
-      borderTopWidth: StyleSheet.hairlineWidth,
-      borderColor: withOpacity(colors.foreground, 0.12),
-      paddingTop: 10,
-    },
-    bottomDockRow: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-      gap: 2,
-    },
-    bottomDockBtn: {
-      flex: 1,
-      height: 54,
-      borderRadius: 10,
-      alignItems: "center",
-      justifyContent: "center",
-      gap: 4,
-      paddingTop: 4,
-    },
-    bottomDockBtnActive: {
-      backgroundColor: withOpacity(colors.foreground, 0.06),
-    },
-    bottomDockLabel: {
-      fontSize: fontSize.xs,
-      lineHeight: 14,
-      color: colors.mutedForeground,
-      fontWeight: fontWeight.medium,
-    },
-    bottomDockLabelActive: {
-      color: colors.primary,
-    },
-    bottomToolbarProgressTrack: {
-      position: "absolute",
-      top: 0,
-      left: 0,
-      right: 0,
-      height: 2,
-      backgroundColor: withOpacity(colors.foreground, 0.08),
-      overflow: "hidden",
-    },
-    bottomToolbarProgressFill: {
-      height: "100%",
-      backgroundColor: withOpacity(colors.foreground, 0.48),
-      borderRadius: 999,
-    },
-    toolbarRow: { flexDirection: "row", alignItems: "center", justifyContent: "space-between" },
-    toolbarBtn: {
-      width: 36,
-      height: 36,
-      borderRadius: 18,
-      alignItems: "center",
-      justifyContent: "center",
-    },
-    toolbarBtnActive: { backgroundColor: `${colors.primary}30` },
-    toolbarBtnDisabled: { opacity: 0.4 },
-    toolbarCenter: { flex: 1, paddingHorizontal: 8, alignItems: "center" },
-    toolbarTitle: {
-      fontSize: fontSize.sm,
-      fontWeight: fontWeight.semibold,
-      color: "#fff",
-      letterSpacing: 0.3,
-    },
-    toolbarChapter: { fontSize: fontSize.xs, color: "rgba(255,255,255,0.5)", marginTop: 1 },
-
-    topInfoBar: {
-      position: "absolute",
-      left: 0,
-      right: 0,
-      paddingHorizontal: 18,
-      paddingVertical: 6,
-      gap: 8,
-    },
-    topInfoRow: {
-      flexDirection: "row",
-      justifyContent: "space-between",
-      alignItems: "center",
-    },
-    topInfoText: {
-      flex: 1,
-      fontSize: fontSize.xs,
-      color: colors.mutedForeground,
-      marginRight: 8,
-    },
-    topInfoPageText: {
-      fontSize: fontSize.xs,
-      color: colors.mutedForeground,
-    },
-    topInfoProgressTrack: {
-      height: 2,
-      borderRadius: 999,
-      overflow: "hidden",
-      backgroundColor: withOpacity(colors.foreground, 0.08),
-    },
-    topInfoProgressFill: {
-      height: "100%",
-      borderRadius: 999,
-      backgroundColor: withOpacity(colors.foreground, 0.42),
-    },
-    bottomInfoBar: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-      position: "absolute",
-      zIndex: 24,
-    },
-    bottomInfoSide: {
-      flexDirection: "row",
-      alignItems: "center",
-      gap: 6,
-    },
-    bottomInfoText: {
-      fontSize: fontSize.xs,
-      color: colors.mutedForeground,
-      fontVariant: ["tabular-nums"],
-    },
-    footerSliderRow: { flexDirection: "row", alignItems: "center", gap: 6 },
-    footerNavBtn: {
-      width: 36,
-      height: 36,
-      borderRadius: 18,
-      alignItems: "center",
-      justifyContent: "center",
-      backgroundColor: "rgba(255,255,255,0.1)",
-    },
-    sliderWrap: { flex: 1, justifyContent: "center", paddingVertical: 4 },
-    sliderTrack: {
-      height: 3,
-      backgroundColor: "rgba(255,255,255,0.15)",
-      borderRadius: 1.5,
-      overflow: "hidden",
-    },
-    sliderFill: { height: "100%", backgroundColor: colors.primary, borderRadius: 1.5 },
-
-    thinProgressWrap: {
-      position: "absolute",
-      left: 0,
-      right: 0,
-      height: 2,
-      backgroundColor: "rgba(255,255,255,0.05)",
-      zIndex: 40,
-    },
-    thinProgressFill: { height: "100%", backgroundColor: colors.primary, opacity: 0.8 },
-
-    searchBarWrap: {
-      position: "absolute",
-      top: 0,
-      left: 0,
-      right: 0,
-      backgroundColor: colors.background,
-      borderBottomWidth: 0.5,
-      borderBottomColor: colors.border,
-      zIndex: 40,
-    },
-    searchBarRow: {
-      flexDirection: "row",
-      alignItems: "center",
-      gap: 4,
-      paddingHorizontal: 12,
-      paddingVertical: 8,
-    },
-    searchInputWrap: {
-      flex: 1,
-      flexDirection: "row",
-      alignItems: "center",
-      backgroundColor: colors.muted,
-      borderRadius: radius.lg,
-      paddingHorizontal: 10,
-      height: 36,
-      gap: 6,
-    },
-    searchInput: { flex: 1, fontSize: fontSize.sm, color: colors.foreground, padding: 0 },
-    searchMetaRow: { flexDirection: "row", alignItems: "center" },
-    searchCount: { fontSize: fontSize.xs, color: colors.mutedForeground },
-    searchNavBtn: {
-      width: 32,
-      height: 32,
-      borderRadius: radius.lg,
-      alignItems: "center",
-      justifyContent: "center",
-    },
-
-    modalBackdrop: { flex: 1, backgroundColor: "rgba(0,0,0,0.4)" },
-    bottomSheet: {
-      backgroundColor: colors.card,
-      borderTopLeftRadius: radius.xxl,
-      borderTopRightRadius: radius.xxl,
-      padding: 16,
-    },
-    sheetHeader: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-      marginBottom: 12,
-    },
-    sheetTitle: {
-      fontSize: fontSize.md,
-      fontWeight: fontWeight.semibold,
-      color: colors.foreground,
-    },
-    sheetScroll: { maxHeight: SCREEN_HEIGHT * 0.5 },
-    sheetEmpty: {
-      fontSize: fontSize.sm,
-      color: colors.mutedForeground,
-      textAlign: "center",
-      paddingVertical: 32,
-    },
-
-    tocTabBar: {
-      flexDirection: "row" as const,
-      alignItems: "center" as const,
-      gap: 4,
-    },
-    tocTab: {
-      flexDirection: "row" as const,
-      alignItems: "center" as const,
-      gap: 4,
-      paddingHorizontal: 8,
-      paddingVertical: 4,
-      borderRadius: radius.md,
-    },
-    tocTabText: {
-      fontSize: fontSize.sm,
-      fontWeight: fontWeight.medium,
-    },
-
-    settingRow: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-      paddingVertical: 12,
-      borderBottomWidth: 0.5,
-      borderBottomColor: colors.border,
-    },
-    settingRowMultiline: {
-      alignItems: "flex-start",
-      gap: 12,
-    },
-    settingLabelBlock: {
-      flex: 1,
-      minWidth: 0,
-      paddingRight: 12,
-      gap: 4,
-    },
-    settingLabel: { fontSize: fontSize.sm, color: colors.mutedForeground },
-    settingHint: {
-      fontSize: fontSize.xs,
-      lineHeight: 16,
-      color: withOpacity(colors.mutedForeground, 0.82),
-    },
-    settingControl: { flexDirection: "row", alignItems: "center", gap: 12 },
-    stepBtn: {
-      width: 32,
-      height: 32,
-      borderRadius: radius.lg,
-      backgroundColor: colors.muted,
-      alignItems: "center",
-      justifyContent: "center",
-    },
-    stepBtnText: { fontSize: fontSize.sm, fontWeight: fontWeight.medium, color: colors.foreground },
-    settingValue: {
-      fontSize: fontSize.sm,
-      fontWeight: fontWeight.medium,
-      color: colors.foreground,
-      minWidth: 32,
-      textAlign: "center",
-    },
-    settingToggleBtn: {
-      minWidth: 68,
-      paddingHorizontal: 12,
-      paddingVertical: 8,
-      borderRadius: radius.lg,
-      backgroundColor: colors.muted,
-      alignItems: "center",
-      justifyContent: "center",
-    },
-    settingToggleBtnActive: {
-      backgroundColor: `${colors.primary}18`,
-    },
-    settingToggleText: {
-      fontSize: fontSize.xs,
-      color: colors.mutedForeground,
-      fontWeight: fontWeight.medium,
-    },
-    settingToggleTextActive: {
-      color: colors.primary,
-    },
-    themeScroll: { maxWidth: 220 },
-    themeRow: { flexDirection: "row", gap: 6 },
-    themeBtn: {
-      paddingHorizontal: 12,
-      paddingVertical: 6,
-      borderRadius: radius.lg,
-      backgroundColor: colors.muted,
-    },
-    themeBtnActive: { backgroundColor: colors.primary },
-    themeBtnText: {
-      fontSize: fontSize.xs,
-      fontWeight: fontWeight.medium,
-      color: colors.mutedForeground,
-    },
-    themeBtnTextActive: { color: colors.primaryForeground },
-    viewModeRow: { flexDirection: "row", gap: 8 },
-    viewModeBtn: {
-      paddingHorizontal: 16,
-      paddingVertical: 6,
-      borderRadius: radius.lg,
-      backgroundColor: colors.muted,
-    },
-    viewModeBtnActive: { backgroundColor: colors.primary },
-    viewModeBtnText: {
-      fontSize: fontSize.xs,
-      fontWeight: fontWeight.medium,
-      color: colors.mutedForeground,
-    },
-    viewModeBtnTextActive: { color: colors.primaryForeground },
-
-    notebookPlaceholder: {
-      alignItems: "center",
-      justifyContent: "center",
-      paddingVertical: 48,
-      gap: 12,
-    },
-    notebookPlaceholderText: {
-      fontSize: fontSize.sm,
-      color: colors.mutedForeground,
-      textAlign: "center",
-      paddingHorizontal: 32,
-    },
-
-    highlightItem: {
-      flexDirection: "row",
-      alignItems: "flex-start",
-      gap: 8,
-      paddingVertical: 8,
-      borderBottomWidth: 0.5,
-      borderBottomColor: colors.border,
-    },
-    highlightColorDot: { width: 8, height: 8, borderRadius: 4, marginTop: 4 },
-    highlightContent: { flex: 1 },
-    highlightText: { fontSize: fontSize.sm, color: colors.foreground, lineHeight: 18 },
-    highlightNote: { fontSize: fontSize.xs, color: colors.mutedForeground, marginTop: 4 },
-
-    bookmarkItem: {
-      flexDirection: "row" as const,
-      alignItems: "center" as const,
-      gap: 10,
-      paddingVertical: 12,
-      paddingHorizontal: 12,
-      borderBottomWidth: 0.5,
-      borderBottomColor: colors.border,
-    },
-    bookmarkContent: {
-      flex: 1,
-      minWidth: 0,
-    },
-    bookmarkLabel: {
-      fontSize: fontSize.sm,
-      fontWeight: fontWeight.medium,
-      lineHeight: 20,
-    },
-    bookmarkSnippet: {
-      fontSize: fontSize.xs,
-      lineHeight: 16,
-      marginTop: 2,
-      opacity: 0.7,
-    },
-    bookmarkDate: {
-      fontSize: fontSize.xs,
-      marginTop: 3,
-      opacity: 0.6,
-    },
-    bookmarkDeleteBtn: {
-      padding: 6,
-      borderRadius: radius.md,
-      opacity: 0.5,
-    },
-
-    noteViewOverlay: {
-      flex: 1,
-      justifyContent: "flex-end",
-      backgroundColor: "rgba(0,0,0,0.5)",
-    },
-    noteViewModal: {
-      backgroundColor: colors.card,
-      borderTopLeftRadius: radius.xxl,
-      borderTopRightRadius: radius.xxl,
-      padding: 16,
-      maxHeight: SCREEN_HEIGHT * 0.75,
-    },
-    noteViewHeader: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-      marginBottom: 12,
-    },
-    noteViewTitle: {
-      fontSize: fontSize.md,
-      fontWeight: fontWeight.semibold,
-      color: colors.foreground,
-    },
-    noteViewCloseBtn: {
-      width: 32,
-      height: 32,
-      borderRadius: 16,
-      alignItems: "center",
-      justifyContent: "center",
-      backgroundColor: colors.muted,
-    },
-    noteViewQuote: {
-      fontSize: fontSize.sm,
-      color: colors.mutedForeground,
-      marginBottom: 12,
-      fontStyle: "italic",
-      lineHeight: 20,
-      paddingHorizontal: 8,
-      borderLeftWidth: 2,
-      borderLeftColor: colors.primary,
-    },
-    noteViewBody: {
-      maxHeight: SCREEN_HEIGHT * 0.35,
-      backgroundColor: colors.muted,
-      borderRadius: radius.lg,
-      paddingHorizontal: 12,
-      paddingVertical: 8,
-    },
-    noteViewEditorContainer: {
-      height: 200,
-      borderRadius: radius.lg,
-      borderWidth: 1,
-      borderColor: colors.border,
-      overflow: "hidden",
-    },
-    noteViewActions: {
-      flexDirection: "row",
-      justifyContent: "flex-end",
-      gap: 8,
-      marginTop: 12,
-    },
-    noteViewEditBtn: {
-      flexDirection: "row",
-      alignItems: "center",
-      gap: 4,
-      backgroundColor: colors.primary,
-      paddingHorizontal: 12,
-      paddingVertical: 8,
-      borderRadius: radius.lg,
-    },
-    noteViewEditText: {
-      fontSize: fontSize.sm,
-      fontWeight: fontWeight.medium,
-      color: colors.primaryForeground,
-    },
-    noteViewCancelBtn: {
-      flexDirection: "row",
-      alignItems: "center",
-      gap: 4,
-      paddingHorizontal: 12,
-      paddingVertical: 8,
-      borderRadius: radius.lg,
-    },
-    noteViewCancelText: {
-      fontSize: fontSize.sm,
-      color: colors.mutedForeground,
-    },
-    noteViewSaveBtn: {
-      flexDirection: "row",
-      alignItems: "center",
-      gap: 4,
-      backgroundColor: colors.primary,
-      paddingHorizontal: 12,
-      paddingVertical: 8,
-      borderRadius: radius.lg,
-    },
-    noteViewSaveText: {
-      fontSize: fontSize.sm,
-      fontWeight: fontWeight.medium,
-      color: colors.primaryForeground,
-    },
-
-    noteTooltip: {
-      position: "absolute",
-      width: 300,
-      maxHeight: 200,
-      backgroundColor: "rgba(15, 23, 42, 0.95)",
-      borderRadius: radius.lg,
-      padding: 12,
-      shadowColor: "#000",
-      shadowOffset: { width: 0, height: 8 },
-      shadowOpacity: 0.3,
-      shadowRadius: 16,
-      elevation: 12,
-      borderWidth: 1,
-      borderColor: "rgba(100, 116, 139, 0.3)",
-      zIndex: 90,
-    },
-    noteTooltipContent: {
-      maxHeight: 140,
-      overflow: "hidden",
-    },
-  });
+export const makeStyles = (colors: ThemeColors) => ({
+  ...makeToolbarStyles(colors),
+  ...makeSheetStyles(colors),
+  ...makeNoteStyles(colors),
+});
